requirements.txt
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy
pandas
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
six==1.16.0
spotipy==2.23.0
tzdata==2024.1
urllib3
boto3
apache-airflow-providers-amazon